# intergalactic-market/src/main/resources/api-spec/product-api-spec.yml
openapi: 3.0.4
info:
  title: Product API
  description: Product API for the Intergalactic Marketplace
  version: 1.0.0

tags:
  - name: Product
    description: CRUD operations for Product entity

paths:
  /api/v1/products:
    get:
      tags:
        - Product
      summary: Get the list of all products
      responses:
        '200':
          description: A list of all products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListDTO'
    post:
      tags:
        - Product
      summary: Create a new product
      requestBody:
        description: Information about product to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '201':
          description: Successful product creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductEntry'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{id}:
    get:
      tags:
        - Product
      summary: Get a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the product to retrieve
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductEntry'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Product
      summary: Update product information
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the product to update
      requestBody:
        description: The product information to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductEntry'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Product
      summary: Delete a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the product to delete
      responses:
        '204':
          description: Product deleted successfully (No Content)
        '500':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ProductDTO:
      type: 'object'
      required: ["name", "price", "category"]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
          description: Name of the product
        price:
          type: number
          format: double
          exclusiveMinimum: 0
          description: Price of the product
        description:
          type: string
          description: Description of the product
        category:
          type: array
          minItems: 1
          items:
            type: string
          description: CosmicCategory to which product belongs

    ProductListDTO:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductEntry'

    ProductEntry:
      type: 'object'
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the product
        name:
          type: string
          description: Name of the product
        price:
          type: number
          format: double
          description: Price of the product
        description:
          type: string
          description: Description of the product
        category:
          type: array
          items:
            type: string
          description: CosmicCategory to which product belongs

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: A short description of the error
        message:
          type: string
          description: Detailed description of the error
        path:
          type: string
          description: The path of the request that caused the error